refactor: rename from InfoItem to DataItem
Change the naming of the InfoItem to be a DataItem as it makes more
sense when reading it as a bit of data, and also more fits the
DeviceInfo
