+++ app/src/main/java/com/example/deviceinfo/dataitems/ManufacturerData.kt
@@ -1,11 +1,11 @@
-package com.example.deviceinfo.infoitems
+package com.example.deviceinfo.dataitems
 
 import android.os.Build
 
 /**
  * Provides information on the Manufacturer of the current device.
  */
-class ManufacturerData() : InfoItem {
+class ManufacturerData() : DataItem {
     override val label: String get() = "Manufacturer"
     override val value: String get() = Build.MANUFACTURER
 }
